+++ core/components/romanesco/elements/plugins/07_computations/c_global/renderstructureddata.plugin.php
@@ -29,7 +29,7 @@ switch ($modx->event->name) {
         // Cached DOM output already includes structured data
         if ($content) {
             $cacheManager = $modx->getCacheManager();
-            $cacheElementKey = '/dom';
+            $cacheElementKey = '/dom.'. hash('xxh3', $_SERVER['REQUEST_URI']);
             $cacheOptions = [
                 xPDO::OPT_CACHE_KEY => 'resource/' . $modx->resource->getCacheKey()
             ];